backend/requirements.txt
fastapi[standard]
sqlmodel
ollama
python-jose[cryptography]
passlib[bcrypt]